Project 5(11) sprint. Fix for review 2.
