+++ tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from selenium import webdriver
+
+@pytest.fixture(scope="function")
+def driver():
+    driver = webdriver.Chrome()
+    driver.get("https://stellarburgers.nomoreparties.site/")
+    yield driver
+    driver.quit()
