+++ tests/test_3transition_personal_account.py
@@ -1,16 +1,10 @@
-import pytest
-from selenium import webdriver
 from locators import Locators
 
-@pytest.fixture
-def driver():
-    driver = webdriver.Chrome()
-    driver.get("https://stellarburgers.nomoreparties.site/")
-    yield driver
-    driver.quit()
+class TestTransitionPersonalAccount:
 
-def test_login_through_personal_account_button_success(driver):
-    driver.find_element(*Locators.PERSONAL_ACCOUNT_BUTTON).click()
-    expected_url = "https://stellarburgers.nomoreparties.site/login"
-    actual_url = driver.current_url
-    assert actual_url == expected_url, f"Ожидался URL {expected_url}, но получен {actual_url}"
+    # Переход по клику на "Личный кабинет"
+    def test_login_through_personal_account_button_success(self, driver):
+        driver.find_element(*Locators.PERSONAL_ACCOUNT_BUTTON).click()
+        expected_url = "https://stellarburgers.nomoreparties.site/login"
+        actual_url = driver.current_url
+        assert actual_url == expected_url, f"Ожидался URL {expected_url}, но получен {actual_url}"
